requirements.txt
langchain==0.2.16
langchain-core==0.2.39
langchain-community==0.2.17
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
faiss-cpu==1.13.0
sentence-transformers==3.0.1
pymupdf==1.24.10
pdfplumber==0.11.4
groq==0.11.0
streamlit==1.38.0
python-dotenv==1.0.1
Pillow==10.4.0
pytesseract==0.3.13
tf-keras==2.16.0
supermemory==3.37.0
fastapi==0.111.1
uvicorn==0.23.2